Add files via upload
The Alarm Clock is a simple yet functional web application built using vanilla JavaScript, HTML, and CSS. It allows users to easily set multiple alarms, view the current time, and manage their alarms with an intuitive interface. Designed with a clean and minimalistic approach, it offers an efficient way to track time and remind users of important events.

The app updates the current time in real-time and triggers an alert with a sound when the alarm goes off. Users can also delete alarms individually from the list. It’s ideal for anyone looking for a basic alarm system to manage tasks or daily reminders in a web-based environment.